+++ src/cli/commands/backup-restore.ts
@@ -1,6 +1,6 @@
 import { execSync } from 'node:child_process';
 import { $command, $options } from '../commands';
-import { Backup } from '../../db/backup';
+import { Backup } from '../../db-v1/backup';
 import { s } from '../../utils';
 
 const options = $options({
