+++ src/cli/commands/dump-download.ts
@@ -1,6 +1,6 @@
 import { basename } from 'node:path';
 import { $command, $options } from '../commands';
-import { Dump } from '../../db/dump';
+import { Dump } from '../../db-v1/dump';
 import { s } from '../../utils';
 
 const options = $options({
